requirements.txt
awswrangler==2.16.1
black==22.6.0
boto3==1.24.45
flake8==4.0.1
functions-framework==3.1.0
google-cloud-pubsub==2.13.4
google-cloud-secret-manager==2.12.1
google-cloud-storage==2.5.0